feat : set datatype as short for version

+++ src/main/java/com/learning/mfscreener/entities/MFSchemeEntity.java
@@ -1,7 +1,16 @@
-/* Licensed under Apache-2.0 2021-2022. */
+/* Licensed under Apache-2.0 2021-2024. */
 package com.learning.mfscreener.entities;
 
-import jakarta.persistence.*;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import jakarta.persistence.Version;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
@@ -32,7 +41,7 @@ public class MFSchemeEntity extends AuditableEntity<String> implements Serializa
     private String schemeNameAlias;
 
     @Version
-    private Long version;
+    private Short version;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "mf_scheme_type_id")
